requirements.txt
streamlit>=1.38.0
chromadb>=0.5.0
sentence-transformers>=3.0.1
groq>=0.5.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
